Revert "Hard-code bf16, and work around missing triu"
This reverts commit 4fc7fd71466cf7b97b97f2d00effd7772c5e7b58.

+++ olmo/model.py
@@ -503,7 +503,6 @@ class Olmo(nn.Module):
                 ),
                 diagonal=1,
             )
-            att_bias = att_bias.to(torch.bfloat16)
             att_bias.masked_fill_(att_bias == 1, float("-inf"))
             self.register_buffer(
                 "_causal_attention_bias",
@@ -519,17 +518,17 @@ class Olmo(nn.Module):
         if not hasattr(self, "_alibi_attention_bias"):
             # shape: (1, 1, 1, seq_len)
             alibi_bias = torch.arange(
-                1 - self.config.max_sequence_length, 1, dtype=torch.bfloat16, device=self.config.device
+                1 - self.config.max_sequence_length, 1, dtype=torch.float, device=self.config.device
             ).view(1, 1, 1, self.config.max_sequence_length)
 
             # shape: (1, 1, seq_len, seq_len)
             alibi_bias = alibi_bias - torch.arange(
-                1 - self.config.max_sequence_length, 1, dtype=torch.bfloat16, device=self.config.device
+                1 - self.config.max_sequence_length, 1, dtype=torch.float, device=self.config.device
             ).view(1, 1, self.config.max_sequence_length, 1)
             alibi_bias.abs_().mul_(-1)
 
             # shape: (n_heads,)
-            m = torch.arange(1, self.config.n_heads + 1, dtype=torch.bfloat16, device=self.config.device)
+            m = torch.arange(1, self.config.n_heads + 1, dtype=torch.float, device=self.config.device)
             m.mul_(self.config.alibi_bias_max / self.config.n_heads)
 
             # shape: (1, n_heads, seq_len, seq_len)
